+++ tests/specs/references/intersectionType/insideOfTypeArguments.d.ts
@@ -1,0 +1,8 @@
+export type RecordEntryObject = {
+    v: string
+    n: number
+}
+
+export type RecordEntryArrayItem = ReadonlyArray<
+    RecordEntryObject & { i: number }
+>